Created a new design for the required annotation classes. These are partially implemented: the structure is there but more logic is required for ensuring accurate and unique metaids for subjects.
The user interface for composite annotations has been implemented. PhysicalPhenomenon.cpp is a superclass of the Physical* classes and the Editor.cpp takes a shared pointer to one of these.

Figured out why serialisation output format wasn't "pretty": The name of the pretty parser is rdfxml-abbrev - not rdfxml. This also takes care of the rdf:Bag tag problem.

Built support for querying in sparql and whichever languages are supported by librdf.

Tests are a bit of a mess at the moment.